% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollmin.R
\name{rollmin}
\alias{rollmin}
\title{Rolling miniumum}
\usage{
rollmin(x, k, fill = if (na.pad) NA, align = c("center", "left", "right"), ...)
}
\arguments{
\item{x}{an object (representing a series of observations).}

\item{k}{integer width of the rolling window.}

\item{fill}{a three-component vector or list (recycled otherwise)
providing filling values at the left/within/to the right of the
data range. See the ‘fill’ argument of na.fill for details.}

\item{align}{character specifying whether the index of the result
should be left- or right-aligned or centered (default) compared
to the rolling window of observations.}

\item{...}{Further arguments passed to methods.}
}
\value{
An object of the same class as ‘x’ with the rolling min.
}
\description{
Rolling miniumum
}
\author{
David C Sterratt
}
